smsc_mikrotik_numbers/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
psycopg2-binary==2.9.6
requests==2.30.0
urllib3==2.0.2
